taskflow/requirements.txt
asgiref==3.11.1
Django==6.0.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg
gunicorn
psycopg2-binary
PyJWT==2.11.0
sqlparse==0.5.5
